BUG FIX - Structure highlight.js script to only load on tutorial page
Previously when you clicked on a link with a "_blank" target the link would open in a new tab, but the main page would also load up the link.  This is a bug caused by highlight.js. Because of this bug you cannot use target="_blank" on tutorials without problems.
THis http://drupal.org/node/1700620 helped me solve the problem.