Remove iOS double click bug fix js as it causes more bugs
When you scroll from an iphone the ux can is suboptimal because this piece of code will open any link you touch while scrolling. Usually a link only opens if you click it, not just becauase you were scrolling and touched it. I am still trying to find a fix for the double click bug that this piece of js solved.